+++ .travis.yml
@@ -1,0 +1,4 @@
+language: ruby
+rvm:
+  - 2.1.7
+before_install: gem install bundler -v 1.10.6
