--- History.txt
@@ -1,4 +1,0 @@
-=== 1.0.0 / 2009-11-03
-
-* Birthday!
-
