--- Manifest.txt
@@ -1,7 +1,0 @@
-.autotest
-History.txt
-Manifest.txt
-README.rdoc
-Rakefile
-lib/geohash.rb
-test/test_geohash.rb
