+++ lib/ibotta_geohash/version.rb
@@ -1,0 +1,3 @@
+module IbottaGeohash
+  VERSION = "0.2.0"
+end
